Add support for the --disable-boehm-gc flag for systems that lack -lgc.
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@485 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/SDCCast.c
@@ -85,7 +85,7 @@ ast  *newAst (int  type, void *op )
 	ex->opval.val = (value *) op;
 	break ;
     case EX_OP     :
-	ex->opval.op   = (int) op ;
+	ex->opval.op   = (long) op ;
 	break ;
     case EX_LINK   :
 	ex->opval.lnk  = (link *) op;
@@ -100,7 +100,7 @@ ast  *newAst (int  type, void *op )
 /*-----------------------------------------------------------------*/
 /* newNode - creates a new node                                    */
 /*-----------------------------------------------------------------*/
-ast  *newNode ( int op,   ast  *left, ast *right   )
+ast  *newNode ( long op,   ast  *left, ast *right   )
 {
     ast  *ex ;
